requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.2.11
django-taggit==5.0.1
gunicorn==20.1.0
psycopg2==2.9.9
sqlparse==0.4.4
